+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Ensure Active Record queries inside a Sequel transaction are typemapped correctly in postgres adapter (@janko)
+
 * Fix executing Active Record queries inside a Sequel transaction not working in mysql2 adapter (@janko)
 
 ## 1.2.1 (2020-01-10)